Convert Toggle to TS
