+++ src/Toggle/Toggle.tsx
@@ -1,6 +1,6 @@
-import React from 'react';
-import ReactDOM from 'react-dom';
-import { skinnable, props, t } from '../utils';
+import * as React from 'react';
+import * as ReactDOM from 'react-dom';
+import { props, t } from '../utils';
 import cx from '../utils/classnames';
 import { getValueLink } from '../LinkState';
 import { warn } from '../utils/log';
@@ -18,32 +18,48 @@ export const Props = {
   style: t.maybe(t.Object)
 };
 
+export type ToggleProps = {
+  /** the current value (`true` if checked) */
+  value?: boolean,
+  /** callback called when user clicks on the Toggle */
+  onChange?: (value: boolean) => void,
+  /** disable the onClick callback and renders with reduced opacity */
+  disabled?: boolean,
+  /** to be used together with `linkState` */
+  valueLink?: {
+    value?: boolean,
+    requestChange: (value: boolean) => void
+  },
+  /** The size for the Toggle in whatever unit (px, em, rem ...). It will be used to compute `width`, `height` and `border-radius` as follows: `width: size`, `height: size / 2`, `border-radius: size / 2` */
+  size?: number |string,
+  className?: string,
+  style?: React.CSSProperties
+};
+
 /**
  * A nice animated Toggle rendered using only CSS
- * @param value - the current value (`true` if checked)
- * @param onChange - callback called when user clicks on the Toggle
- * @param disabled - disable the onClick callback and renders with reduced opacity
- * @param valueLink - to be used together with `linkState`
- * @param size - The size for the Toggle in whatever unit (px, em, rem ...). It will be used to compute `width`, `height` and `border-radius` as follows: `width: size`, `height: size / 2`, `border-radius: size / 2`
  */
-@skinnable()
 @props(Props)
-export default class Toggle extends React.PureComponent {
+export default class Toggle extends React.PureComponent<ToggleProps> {
 
   componentDidMount() {
     this.updateCheckbox(this.props);
   }
 
-  updateCheckbox = (props) => {
+  componentWillReceiveProps(nextProps: ToggleProps) {
+    this.updateCheckbox(nextProps);
+  }
+
+  updateCheckbox = (props: ToggleProps) => {
     const { value } = getValueLink(this, props);
-    const checkboxNode = ReactDOM.findDOMNode(this.refs.checkbox);
+    const checkboxNode = ReactDOM.findDOMNode<HTMLInputElement>(this.refs.checkbox);
     checkboxNode.checked = value;
   };
 
-  getHalfSize(size) {
+  getHalfSize(size: string | number) {
     if (t.String.is(size)) {
       const unitMatch = (/[a-z]+$/).exec(size); // only match characters at the end
-      const number = parseFloat(size, 10);
+      const number = parseFloat(size);
       const unit = unitMatch ? unitMatch[0] : '';
       if (isFinite(number)) { // we can still get NaN from parseFloat
         return `${number / 2}${unit}`;
@@ -61,36 +77,27 @@ export default class Toggle extends React.PureComponent {
     requestChange(!value);
   };
 
-  getLocals() {
+  render() {
     const {
-      props: { className, size, style, disabled },
+      props: { className: _className, size, style, disabled },
       onButtonClick
     } = this;
-    const { value } = getValueLink(this, this.props);
-    return {
-      style,
-      value,
-      buttonProps: {
-        onClick: disabled ? null : onButtonClick,
-        style: size ?
-          { width: size, height: this.getHalfSize(size), borderRadius: this.getHalfSize(size) } :
-          undefined
-      },
-      className: cx('toggle', { disabled }, className)
-    };
-  }
 
-  template({ value, className, style, buttonProps }) {
+    const { value } = getValueLink(this, this.props);
+    const buttonProps = {
+      onClick: disabled ? undefined : onButtonClick,
+      style: size ?
+        { width: size, height: this.getHalfSize(size), borderRadius: this.getHalfSize(size) } :
+        undefined
+    };
+    const className = cx('toggle', { disabled }, _className);
+
     return (
       <div {...{ className, style }}>
         <input className='toggle-input' type='checkbox' ref='checkbox' value={value} readOnly />
         <label className='toggle-button' {...buttonProps} />
       </div>
     );
-  }
-
-  componentWillReceiveProps(nextProps) {
-    this.updateCheckbox(nextProps);
   }
 
 }
