--- src/Toggle/index.js
@@ -1,2 +1,0 @@
-import Toggle from './Toggle';
-export default Toggle;
